Fixed bug where Gama Data Handler crashed + better logs
Fixed minor issues where we had missing functions in dummy objects
Also nodes now hold a dummy number for easier debugging.

The main fix was in cases where we had two edges that were the same and
we trying to clean the same edge twice.
We now clean the second edge only if its different then the first.